% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsftheme.R
\name{ucsf_col}
\alias{ucsf_col}
\title{Function to extract official UCSF colors as hex codes.}
\usage{
ucsf_col(...)
}
\arguments{
\item{...}{Character names of UCSF color.}
}
\description{
Executing this function without
any arguments provides a listing of color names and their associated hex codes.
These colors are based on the UCSF Identity Guidelines, available at
\url{https://identity.ucsf.edu/brand-guide/print-colors} and
\url{https://identity.ucsf.edu/brand-guide/digital-colors}.
The base colors are named navy, teal, green, blue, orange.
Accent colors are purple, red, yellow.
Each color has 3 lighter tints, a 70\%, 40\%, and 10\% tint of the main color.
These are names based on the main color, appended with a 3, 2, or 1, respectively.
For example, full strength is called "blue", the 70\% tint is "blue3", the 40\% tint
is "blue2", and the 10\% tint is "blue1"
}
\examples{
ucsf_col("navy")
ucsf_col("teal")
ucsf_col("green")
ucsf_col("blue")
ucsf_col("orange")
ucsf_col("red")
ucsf_col("purple")
ucsf_col("yellow")
ucsf_col("red3")
}
