% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsftheme.R
\name{ucsf_pal}
\alias{ucsf_pal}
\title{Return function to interpolate a UCSF color palette}
\usage{
ucsf_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ucsf_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a UCSF color palette
}
